Refactor stale-features useEffect into FeatureManager
