+++ react/src/components/FeatureManager/FeatureManager.tsx
@@ -1,6 +1,11 @@
 import React, { useEffect } from 'react';
 import { Dayjs } from 'dayjs';
-import { useCurrentFeaturesStore, useFeatures } from '@hazmapper/hooks';
+import {
+  useCurrentFeaturesStore,
+  useFeatures,
+  KEY_USE_FEATURES,
+} from '@hazmapper/hooks';
+import { useQueryClient } from '@tanstack/react-query';
 
 interface FeatureManagerProps {
   projectId: number;
@@ -12,7 +17,7 @@ interface FeatureManagerProps {
 
 /**
  * FeatureManager is responsible for fetching and managing feature data using React Query (`useFeatures`)
- * and syncing the results with Zustand (`useCurrentFeaturesStore`).
+ * and syncing the results with Zustand (so that `useCurrentFeaturesStore` can be used in different places).
  */
 export const FeatureManager: React.FC<FeatureManagerProps> = ({
   projectId,
@@ -21,7 +26,8 @@ export const FeatureManager: React.FC<FeatureManagerProps> = ({
   endDate,
   toggleDateFilter,
 }) => {
-  const { setCurrentFeatures, setCurrentStatus } =
+  const queryClient = useQueryClient();
+  const { setCurrentFeatures, setCurrentStatus, setReset } =
     useCurrentFeaturesStore.getState();
 
   const { data, isFetching, isPending, isError } = useFeatures({
@@ -32,9 +38,18 @@ export const FeatureManager: React.FC<FeatureManagerProps> = ({
     toggleDateFilter,
   });
 
+  // Clear feature queries and our stored data when changing projects to prevent stale features from
+  // briefly appearing and causing incorrect map bounds/zoom during navigation
+  useEffect(() => {
+    return () => {
+      queryClient.removeQueries({ queryKey: [KEY_USE_FEATURES] });
+      setReset();
+    };
+  }, [projectId, setReset, queryClient]);
+
   /**
    * Syncs the current query state from React Query to the Zustand store (`useCurrentFeatures`).
-   * See useCurrentFeatures for more invo
+   * See useCurrentFeatures for more information.
    */
   useEffect(() => {
     if (data) {
